+++ Sources/SuperwallKit/Identity/UserAttributes.swift
@@ -8,10 +8,12 @@
 import Foundation
 
 public extension Superwall {
-  /// Set user attributes for use in your paywalls and the dashboard.
+  /// Sets user attributes for use in paywalls and the Superwall dashboard.
   ///
-  /// If the existing user attributes dictionary already has a value for a given property, the old value is overwritten. Other existing properties will not be affected.
-  /// Useful for analytics and conditional paywall rules you may define in the Superwall Dashboard. They should **not** be used as a source of truth for sensitive information.
+  /// If the existing user attributes dictionary already has a value for a given property, the old
+  /// value is overwritten. Existing properties will not be affected.
+  /// Useful for analytics and conditional paywall rules you may define in the Superwall Dashboard.
+  /// They should **not** be used as a source of truth for sensitive information.
   ///
   /// Here's how you might set user attributes after retrieving your user's data:
   ///  ```swift
@@ -26,26 +28,32 @@ public extension Superwall {
   ///  ```
   /// See <doc:SettingUserAttributes> for more.
   ///
-  ///
-  /// - Parameter custom: A `[String: Any?]` map used to describe any custom attributes you'd like to store to the user. Remember, keys begining with `$` are reserved for Superwall and will be dropped. Values can be any JSON encodable value, URLs or Dates. Arrays and dictionaries as values are not supported at this time, and will be dropped.
+  /// - Parameter attributes: A `[String: Any?]` dictionary used to describe any custom
+  /// attributes you'd like to store for the user. Values can be any JSON encodable value, `URL`s or `Date`s.
+  /// Note: Keys beginning with `$` are reserved for Superwall and will be dropped. Arrays and dictionaries
+  /// as values are not supported at this time, and will be dropped.
   func setUserAttributes(_ attributes: [String: Any?]) {
     mergeAttributes(attributes)
   }
 
-  /// The Objective-C method for setting user attributes for use in your paywalls and the dashboard. **Note**: Please use ``SuperwallKit/Superwall/setUserAttributes(_:)`` if you're using Swift.
+  /// The Objective-C method for setting user attributes for use in your paywalls and the dashboard.
   ///
-  /// If the existing user attributes dictionary already has a value for a given property, the old value is overwritten. Other existing properties will not be affected.
-  /// Useful for analytics and conditional paywall rules you may define in the web dashboard. They should not be used as a source of truth for sensitive information.
+  /// If the existing user attributes dictionary already has a value for a given property, the old
+  /// value is overwritten. Existing properties will not be affected.
+  /// Useful for analytics and conditional paywall rules you may define in the Superwall Dashboard.
+  /// They should **not** be used as a source of truth for sensitive information.
   ///
-  /// - Parameter attributes: A `NSDictionary` used to describe user attributes and any custom attributes you'd like to store to the user. Remember, keys begining with `$` are reserved for Superwall and will be dropped. Values can be any JSON encodable value, URLs or Dates. Arrays and dictionaries as values are not supported at this time, and will be dropped.
+  /// Here's how you might set user attributes after retrieving your user's data:
   ///
-  /// We make our best effort to pick out "known" user attributes and set them to our names. For exampe `{"first_name": "..." }` and `{"firstName": "..."}` will both be translated into `$first_name` for use in Superwall where we require a first name.
-  ///
-  ///  Example:
   ///  ```
   ///  NSDictionary *userAttributes = @{ key : value, key2 : value2};
-  ///  [Superwall setUserAttributesDictionary: userAttributes];
+  ///  [[Superwall sharedInstance] setUserAttributesDictionary: userAttributes];
   ///  ```
+  ///
+  /// - Parameter attributes: An `NSDictionary` used to describe any custom
+  /// attributes you'd like to store for the user. Values can be any JSON encodable value, `URL`s or `Date`s.
+  /// Note: Keys beginning with `$` are reserved for Superwall and will be dropped. Arrays and dictionaries
+  /// as values are not supported at this time, and will be dropped.
   @available(swift, obsoleted: 1.0)
   @objc func setUserAttributesDictionary(_ attributes: NSDictionary) {
     if let anyAttributes = attributes as? [String: Any] {
@@ -57,11 +65,11 @@ public extension Superwall {
     }
   }
 
-  /// The Objective-C method for removing user attributes for use in your paywalls and the dashboard. **Note**: Please use ``SuperwallKit/Superwall/setUserAttributes(_:)`` in the form `["someKey": nil]` if you're using Swift.
+  /// The Objective-C method for removing user attributes for use in your paywalls and the dashboard.
   ///
   ///  Example:
   ///  ```
-  ///  [Superwall removeUserAttributes:@[@"key1", @"key2"]];
+  ///  [[Superwall sharedInstance] removeUserAttributes:@[@"key1", @"key2"]];
   ///  ```
   ///
   /// - Parameter keys: An array containing the keys you wish to remove from the user attributes dictionary.